Move ls command into own file
